Rename noop processor to filter
