+++ plugins/processors/all/filter.go
@@ -1,0 +1,5 @@
+//go:build !custom || processors || processors.filter
+
+package all
+
+import _ "github.com/influxdata/telegraf/plugins/processors/filter" // register plugin
