--- plugins/processors/all/noop.go
@@ -1,5 +1,0 @@
-//go:build !custom || processors || processors.noop
-
-package all
-
-import _ "github.com/influxdata/telegraf/plugins/processors/noop" // register plugin
