+++ plugins/processors/filter/README.md
@@ -1,6 +1,6 @@
-# Noop Processor Plugin
+# Filter Processor Plugin
 
-The noop processor plugin does nothing to metrics. Instead it can be used to
+The filter processor plugin does nothing to metrics. Instead it can be used to
 apply the global configuration options after other processing. Global config
 options like tagpass, fieldpass, and others are applied before a processor,
 aggregator, or output. As such a user might want to apply these after doing
@@ -54,7 +54,7 @@ longer need that tag for additional processing:
   order = 1
   ...
 
-[[processors.noop]]
+[[processors.filter]]
   order = 2
   tagexclude = ["useless_tag"]
 ```
