+++ plugins/processors/filter/sample.conf
@@ -1,5 +1,5 @@
-# Do nothing processor
-[[processors.noop]]
+# Filter metrics by the given criteria
+[[processors.filter]]
 
 ## Metric Filtering
 ## The following options provide mechanisms to include or exclude entire
